Changes to class and adds unit tests
