--- settings.py
@@ -1,4 +1,0 @@
-import os
-
-# Discord API Settings
-DISCORD_WEBHOOK_URL = str(os.environ.get('DISCORD_WEBHOOK_URL'))
